Update quota service to send the container logs directly to the quotaService without passing by the database
